Split Warp config: Linux/Windows shortcuts in warp.py, Mac in warp_mac.py
Adds support for Warp on Linux and Windows (previously only Mac). The
existing warp.py is moved to warp_mac.py and scoped with `os: mac` so
its Mac-specific actions (cmd-{n} tab jumps, alt-arrow word navigation)
no longer leak into other OSes. The new warp.py declares mod.apps.warp
for Linux (app.name) and Windows (app.name + app.exe per P04), and
overrides app.tab_open / app.tab_close (ctrl-shift-t / ctrl-shift-w)
plus user.tab_jump / user.tab_final (ctrl-{n}) to match Warp's
Linux/Windows defaults.
