+++ apps/warp/warp.py
@@ -3,8 +3,13 @@ from talon import Context, Module, actions
 mod = Module()
 
 mod.apps.warp = """
-os: mac
-and app.bundle: dev.warp.Warp-Stable
+os: linux
+and app.name: warp
+"""
+mod.apps.warp = """
+os: windows
+and app.name: Warp
+and app.exe: warp.exe
 """
 
 ctx = Context()
@@ -13,20 +18,20 @@ app: warp
 """
 
 
+@ctx.action_class("app")
+class AppActions:
+    def tab_open():
+        actions.key("ctrl-shift-t")
+
+    def tab_close():
+        actions.key("ctrl-shift-w")
+
+
 @ctx.action_class("user")
 class UserActions:
     def tab_jump(number: int):
         if number < 9:
-            actions.key(f"cmd-{number}")
+            actions.key(f"ctrl-{number}")
 
     def tab_final():
-        actions.key("cmd-9")
-
-
-@ctx.action_class("edit")
-class EditActions:
-    def word_left():
-        actions.key("alt-left")
-
-    def word_right():
-        actions.key("alt-right")
+        actions.key("ctrl-9")
